r2attunet_train/requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
flake8==6.1.0
fonttools==4.42.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.25.2
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.1.0
requests==2.31.0
setuptools==68.0.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
wheel==0.38.4
zipp==3.16.2
